add teams cleanup
